pyproject.toml
[tool.poetry]
name = "minimal-rag"
version = "0.1.0"
description = "Minimal RAG baseline system for research on memory behavior"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
tiktoken = "^0.5.0"
numpy = "^1.24.0"
sentence-transformers = "^2.2.0"
faiss-cpu = "^1.7.4"
google-generativeai = "^0.8.0"
PyPDF2 = "^3.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

